moved filter method to RSVP Email class
