+++ src/Tickets/Emails/Hooks.php
@@ -19,6 +19,7 @@ namespace TEC\Tickets\Emails;
 
 use \tad_DI52_ServiceProvider;
 use TEC\Tickets\Emails\Admin\Emails_Tab;
+use TEC\Tickets\Emails\Email\RSVP;
 use Tribe__Tickets__Tickets as Tickets_Module;
 
 /**
@@ -75,7 +76,7 @@ class Hooks extends tad_DI52_ServiceProvider {
 		add_filter( 'tec_tickets_send_rsvp_non_attendance_confirmation_pre', [ $this, 'filter_hijack_legacy_send_rsvp_non_attendance_confirmation' ], 20, 4 );
 
 		// skip saving hidden fields for RSVP emails.
-		add_filter( 'tribe_settings_fields', [ $this, 'filter_rsvp_fields_for_validation' ], 90, 2 );
+		add_filter( 'tribe_settings_fields', [ $this, 'filter_rsvp_fields_before_saving' ], 90, 2 );
 	}
 
 	/**
@@ -88,31 +89,8 @@ class Hooks extends tad_DI52_ServiceProvider {
 	 *
 	 * @return array
 	 */
-	public function filter_rsvp_fields_for_validation( array $fields, $admin_page ): array {
-		$emails_slug = Emails_Tab::$slug;
-
-		if ( \Tribe\Tickets\Admin\Settings::$settings_page_id !== $admin_page
-		     || ! isset( $fields[ $emails_slug ] )
-		     || tribe_get_request_var( 'section', false ) !== 'tec_tickets_emails_rsvp'
-		     || empty( $_POST )
-		) {
-			return $fields;
-		}
-
-		if ( ! tribe_is_truthy( tribe_get_request_var( 'tec-tickets-emails-rsvp-use-ticket-email' ) ) ) {
-			return $fields;
-		}
-
-		// filter out hidden fields from the fields to be saved.
-		$fields[ $emails_slug ] = array_filter(
-			$fields[ $emails_slug ],
-			function ( $field, $key ) {
-				return isset( $_POST[ $key ] ) || $field['type'] === 'toggle';
-			},
-			ARRAY_FILTER_USE_BOTH
-		);
-
-		return $fields;
+	public function filter_rsvp_fields_before_saving( array $fields, $admin_page ): array {
+		return $this->container->make( RSVP::class )->filter_rsvp_fields_before_saving( $fields, $admin_page );
 	}
 
 	/**
